backend/requirements.txt
Flask
flask-cors
easyocr
opencv-python
numpy
pillow